fix: move verify_return_type out of worker.py to fix test import
