+++ python/pyspark/worker.py
@@ -26,21 +26,7 @@ import time
 import inspect
 import itertools
 import json
-from typing import (
-    Any,
-    Callable,
-    Iterable,
-    Iterator,
-    Optional,
-    Tuple,
-    Type,
-    TypeVar,
-    TYPE_CHECKING,
-    Union,
-    overload,
-)
-
-T = TypeVar("T")
+from typing import Any, Callable, Iterable, Iterator, Optional, Tuple, TYPE_CHECKING, Union
 
 if TYPE_CHECKING:
     from pyspark.sql.pandas._typing import GroupedBatch
@@ -88,6 +74,7 @@ from pyspark.sql.pandas.serializers import (
     ArrowStreamArrowUDTFSerializer,
 )
 from pyspark.sql.pandas.types import to_arrow_schema, to_arrow_type
+from pyspark.sql.pandas._verify import verify_return_type
 from pyspark.sql.types import (
     ArrayType,
     BinaryType,
@@ -245,58 +232,6 @@ def report_times(outfile, boot, init, finish, processing_time_ms):
 def chain(f, g):
     """chain two functions together"""
     return lambda *a: g(f(*a))
-
-
-@overload
-def verify_return_type(result: Any, expected_type: Type[T]) -> T: ...
-
-
-@overload
-def verify_return_type(result: Any, expected_type: Any) -> Any: ...
-
-
-def verify_return_type(result: Any, expected_type: Any) -> Any:
-    """
-    Verify a UDF return value against an expected type.
-
-    Returns ``result`` unchanged if ``isinstance(result, expected_type)``.
-    For ``Iterator[T]``, returns a lazy iterator that checks each element
-    against ``T`` on consumption. Raises ``PySparkTypeError`` on mismatch.
-    """
-    if getattr(expected_type, "_name", None) == "Iterator":
-        (element_type,) = expected_type.__args__
-        package = getattr(inspect.getmodule(element_type), "__package__", "")
-        label = f"iterator of {package}.{element_type.__name__}"
-
-        if not isinstance(result, Iterator):
-            raise PySparkTypeError(
-                errorClass="UDF_RETURN_TYPE",
-                messageParameters={"expected": label, "actual": type(result).__name__},
-            )
-
-        def check_element(element: T) -> T:
-            if not isinstance(element, element_type):
-                raise PySparkTypeError(
-                    errorClass="UDF_RETURN_TYPE",
-                    messageParameters={
-                        "expected": label,
-                        "actual": f"iterator of {type(element).__name__}",
-                    },
-                )
-            return element
-
-        return map(check_element, result)
-
-    if not isinstance(result, expected_type):
-        package = getattr(inspect.getmodule(expected_type), "__package__", "")
-        raise PySparkTypeError(
-            errorClass="UDF_RETURN_TYPE",
-            messageParameters={
-                "expected": f"{package}.{expected_type.__name__}",
-                "actual": type(result).__name__,
-            },
-        )
-    return result
 
 
 def verify_result_row_count(result_length: int, expected: int) -> None:
